+++ tests/client/e2eTests/scripts/importEntities.js
@@ -39,7 +39,8 @@
  *           "name": "Core administrator", // User name
  *           "email": "core-admin@veo-labs.com", // User email
  *           "password": "core-admin", // User password
- *           "roles": [ "coreAdmin", "guest" ] // User role reference as defined in roles property
+ *           "roles": [ "coreAdmin", "guest" ], // User role reference as defined in roles property
+ *           "locked": false // Indicates if user can be modified or not
  *         }
  *       },
  *       "applications": { // Define applications to import